Return a 400 error if logging in returns an invalid code
If we get here we have a `code` but the auth server has rejected it,
meaning that the `code` is not actually one the auth server gave us.

This could happen if the code is changed or re-used.  It shouldn't
happen normally (and indeed we've only seen this happen twice in
production).