Redirect related_to_brexit params to the topic
The related_to_brexit checkbox has a fair bit of code to support it. It was
only available on the news-and-communications finder, and replicated what
you'd get if you selected the Brexit topic.

We have very few users of this, and those that do can be supported by redirecting
them to use the Brexit topic instead.