Submit JWTs to the new account endpoint & pass ID in ?state
To move us towards a more standard OAuth journey, we now have a
work-around to POSTing a signed JWT to the auth endpoint.

Instead, we submit the JWT to an API endpoint, protected by an access
token.  This means we don't need to sign the JWT or pass the OAuth
uid.  We then pass the JWT's ID in the standard OAuth state param.

Note that this is only a temporary thing: ultimately this logic will
be moved to the new account-api app we will be creating, and when the
Transition Checker is switched off, the JWT thing will go away for
good.  The reason for implementing it here and now is to gradually
test the migration to the new account-api approach.

The redirect_path in the `save_results_sign_up` page goes to the
`save_results_confirm` page because if the user is already logged in
we want them redirected back there, so they can review the old and new
results.