Throw an OAuthFailure from tokens! method
This ensures that OAuthFailure is the only error exposed outside of
the OidcClient instance.

An exception snuck through the brexit checker controller because it
was doing this:

```ruby
def save_results_sign_up
  tokens = Services.oidc.tokens!
  # ...
rescue OidcClient::OAuthFailure
  # ...
end
```

But that `rescue` block didn't catch the error, as `tokens!` was
raising an `Rack::OAuth2::Client::Error`.