Remove finder-frontend session cookie
This is used to implement CSRF protection, but isn't actually needed
in this case:

- `save_results_sign_up` - this is just a fancy redirect, the user
will end up on the accounts domain at the login screen.  And in fact
the `confirm_email_signup` method, which is also a redirect behind a
POST request, also disables the CSRF protection.

- `save_results_apply` - this updates the user's saved results, so
this being vulnerable to CSRF would be bad.  However, it's not.  Our
session cookie[1] has `samesite=strict`, which means it's not sent on
cross-site requests.  This means that a cross-site POST request to
this method will not log the user out.

So neither place which was using the CSRF protection actually needed
it.  By disabling it, we can remove this session cookie.  Since we
enabled it in the first place, that's good.

[1] https://github.com/alphagov/govuk-cdn-config/blob/master/vcl_templates/www.vcl.erb#L452