Switch to the account-api auth endpoints
We no longer need to mess around with JWTs!  The sessions controller
is much simpler!

The set_account_session_header method in AccountConcern has become a
little more complex, as there is still code using the tokens directly
for managing the attribute & email subscription.  But when that is
gone, we can simplify the concern to treating the
govuk_account_session as a black box.