Use question description consistently
We currently use a mix of hint elements and 'description' elements (plain paragraphs) to provide additional information for questions. This updates two questions to align the approach. This means using hint elements only when for checkboxes where it is essential to remind people to select all answers that apply. For the rest of the questions, we use description to avoid overloading screen readers with information at each interaction with the form elements.