Remove QA pattern
This is in prep for rewriting https://www.gov.uk/uk-nationals-living-eu as
a standard GOV.UK page.

This is being tracked as tech debt, as it's a QA pattern without a use right
now: https://trello.com/c/FTfMsw1y/163-finder-frontend-has-two-static-qa-patterns-for-brexit

This is not the Brexit checker!

We'll have to [unpublish and redirect the content item](https://github.com/alphagov/search-api/blob/master/lib/special_route_publisher.rb#L102-L107)
before we merge this.