Replace SessionController routes with frontend URL
We have moved the SessionController from finder-frontend. Previously
code was within finder-frontend to allow for a GOV.UK Accounts discovery
to apply to only the Brexit Checker.

In a new phase of the discovery we are expanding the login to a
GOV.UK-wide account. This now fits better in `frontend` where it can
apply globally, instead of in `finder-frontend` which has more specific
applications.

We begin by removing the `/transtion-check/...` session routes for
logging in, logging out and the callback.

These routes are now handled by router, reading from a content item that
was published with `special-route-publisher` ([see PR#152][1]). This
sends the user to the frontend application where it's own routes file
will take over.

Initially I tried having no replacement routes here, however this caused
issues when rails went looking for these routes in tests and via
helpers. I also tried using rail's [direct routes][2]. Hoping that this
would allow the usual sign_in_path and sign_in_url helpers to appear.

In the end I chose to mimic the approach taken on [routes.rb:L39][3],
using a proc to stub out a controller to appease rails, whilst knowing
that router will step in to divert traffic to the fronted microservice
on production.

[1]: https://github.com/alphagov/special-route-publisher/pull/152
[2]: https://guides.rubyonrails.org/routing.html#direct-routes
[3]: https://github.com/alphagov/finder-frontend/pull/2472/commits/6266e37a276c552317767b89f86d21ae6403b54f#diff-959bc9abc46a55332bb64d5155a79323afa75a50ec1a2137ddd22d926f62c6c5R39