Remove checks for account enabled and available
Previously the accounts saw the Brexit checker (on GOV.UK
infrastructure) communicating with the account manager application
directly. The account manager applciation was new, somewhat unproven and
hosted on PaaS.

So we thought it would be prudent to provide a couple of safeguards. One
would allow us to unset a feature flag which would disable the account
service. The other safeguard would ping the account manager to check if
it is still there.

If either of these two checks failed, we would fall back to a service
unavailable page.

[You can find out more about this piece of work here][1]

Fast forward to now, we have started moving some of this infrastructure
onto GOV.UK proper. The accounts-api now handles requests to the
accounts manager and frontend will route user journeys.

The Account manager itself is now slightly more proven, and has better
monitoring. So the initial concern about it falling over or going away
is less of a pressing concern.

We should review the availbility problem and consider a standard way of
allowing the account-api to inform GOV.UK apps if sometheing is wrong
and fail gracefully.

However until then, we can remove these older safegards and simplify a
great deal of code.

[1]: https://trello.com/c/LaSdrZyI/459-feature-turn-off-accounts