Enable caching unless user is logged in
Previously we had a setting to enable caching for the results page
unless the accounts were enabled. We've now removed the "enabled /
disabled" checks.

We are setting a Vary response header in the `set_account_variant`, so
one user will never see the page cached for another.

Given that, and the demise of account enabled, we can now roll the
results page into the standard caching before_action