Rename END SESSION header
Specifically this is now very similar code to what is in [frontend][1]
and we are bringing it further into line here.

This is a header read from the response to end the session so lets give
it a name along those lines.

To Squash into a single commit after review.

[1]: https://github.com/alphagov/frontend/blob/master/app/controllers/concerns/account_concern.rb#L8