Fix minor filter bug when switching breakpoints
Fixes the issue described in #2446 wherein the filters disappear on
desktop when the browser window is resized from desktop to smaller
breakpoints and back.
Instead of programatically applying inline styles to the facets, apply a
class to toggle visibility on/off. This way we can ensure that the
filters are always visible on desktop using CSS media queries.