Reauthenticate, but don't log out, on 403
If the user gets a 403 response, then they are logged in but not at a
high enough level of authentication.  The straightforward way to
handle this would be to treat it the same as a 401: end the user's
session and redirect them to log in.

But that's not a great user experience.  If they decide they don't
want to set up MFA, and come back, they're no longer able to see their
saved pages any more, because they've been logged out.  Better to keep
the user's session intact, but to put them in a "must_reauthenticate?"
bucket.

"must_reauthenticate?" is essentially the same as "!logged_in?" but
doesn't require actually ending the session.  This is probably a
pattern we'll find useful in other frontend apps.