Remove section about fixed filters
I was going to move this to its own manual, but after a quick read
of it I found I struggled to understand:

- What it's for.
- What it means.

Trying some of the links doesn't help - they don't seem to match the
description in this section. Since this was written back in 2016 [1]
and much has changed since then, I think it's better to delete it,
noting that it was always pointing off to Search API docs anyway.

[1]: https://github.com/alphagov/finder-frontend/pull/263