Move Finder development section into own doc
This isn't necessary for someone to get started with the app. I've
removed the part about local Finder development, since the env var
referenced there isn't present in the codebase, which suggests it
no longer works as described. While the resulting doc is quite short
and vague (especially the third bullet), it's a starting point for
someone to elaborate on in future.