If the user has no saved results, skip the comparison table
This is the first step of untangling the special-case sign up
behaviour for the Transition Checker: we need to be able to send the
user to the same place after authenticating, whether they log in to an
existing account or create a new one.

We need to gracefully handle the case where the user has no current
saved results.  To do that, skip the comparison table (because
everything is new), and go straight on to the email sign up page.