Skip the redirect via /sign-in to authenticate a user
The /sign-in endpoint is there because we need something which the
"sign in" link in the header can go to.  But all it's going is
validating the parameters and then calling the account-api to generate
the *real* sign-in URL.

We don't need to redirect via /sign-in from finder-frontend, because
we can hit account-api directly to generate the URL.  This removes one
redirect from the user's journey, making signing in very slightly
faster.

When this change has been deployed, we can remove the `state_id`
parameter from the frontend `SessionsController` entirely.