Add a CustomEvent polyfill
This is so that Internet Explorer (9+) can handle custom JS events. This
is done to avoid us having to write conditionals each time we want to
fire an event.

As this is a pretty modest polyfill it doesn't strike me that this needs
an immediate wider concern as to whether this belongs in this app or in
static.