Use a CustomEvent when changing search facets
Using a jQuery event meant that only jQuery event handlers would catch
the event. As we are changing the event handlers for these elements to
be native JS ones we need to have events that can be caught by them, as
such CustomEvent is a good choice as both jQuery and native event
handlers will catch them.

Accessing a property on a custom event is a little different with jQuery
which sets it directly on the object, whereas with CustomEvent there is
a field (detail) that is designated for this. I've had to add a
conditional in the code to cater for this. I've added some tests around
it too.