Make Dockerfile production focused
An overlooked aspect of this Dockerfile is that it sets a dev
environment and installs dev/test gems but never actually uses them or has
enough dependencies to run the test suite.

This corrects that by setting the Dockerfile to production environments
and using the new syntax for bundler to avoid warnings.