Customise metadata descriptions for brexit child taxons
Finder frontend uses the description field of a content item to
produces the summary text (aka metadata) underneath each search result.
Because the brexit child taxons are using the description field to populate a
navigation link on the detailed guide page instead, we need to put this text
somewhere else.
An alternative would be to use the description field to store the summary text in
the conventional manner, and hard code the the navigation link in the
detailed guide template instead. However the idea of using a detailed guide page for
the brexit child taxons is to reduce developer input required to make content changes
and bring the pages closer to a standard govuk format. I reckon it's a safer bet that the nav
link will need updating more frequently than the search result metadata, so having the description
field hardcoded in finder frontend is possibly the least bad thing.