Replace `on` method with an eventListener
I've spoken to other members of the team about adding a guard clause to the start of this method. This would have guarded against the possibility that the filter being removed wouldn't have had a corresponding object to remove and would have caused errors. We have since found that there are hidden objects within the window that are required to make this page run. As such we've decided that the best decision is to not add the clause to make sure we can catch errors as they occour. Adding the guard would have meant that the page wouldn't work as expected if the element didn't exist and we'd never know.