Fix tests broken by changes to clearFacet methods
This ensures that all of the tests are working against variables that actually exist. Perviously they were passing in part because jQuery was not falling over when handed undefined variables. They were undefined because of some collisions and confusions around item values. These have been added and amended to betterreflect the items the tests need to pass.