Notifications for these changes
Scenario: 3 (UPDATE ACTION MAJOR)
Change note:
The deadline for applying to the EU Settlement Scheme was 30 June for most people. There are some exceptions.

Scenario: 3 (UPDATE ACTION MAJOR)
Change note:
The deadline for applying to the EU Settlement Scheme was 30 June for most people. There are some exceptions.

Scenario: 1 (NEW ACTION)
The deadline for applying to the EU Settlement Scheme was 30 June for most people. There are some exceptions.