Do not send initial attributes to Account Manager to auth
We originally went for this approach of passing attributes along to
the Account Manager so that we could still save someone's results and
set up their email subscription if they didn't return to GOV.UK after
signing up.

We have now changed the design of the post-registration confirmation
page to make it clearer that the user does need to return to
GOV.UK (and put a big green button on it), and so we no longer think
that supporting users who don't return is worth the complication.  SO
we're removing it entirely.