Fix duplicated banner on transition results page.
There is an edge case on the transition results page, where if the user
must not reauthenticate, and their results differ from the ones already
stored in their account, the "These results are different from the ones
stored in your account" message appears twice.

This removes the duplicated one, so that the message will only appear
once in those specific conditions.