Remove enable-aria-controls JS module
This is being removed as this code no longer does anything. It
previously used to set the aria-controls attribute dynamically on
elements and was replaced in 2018 [1] [2] (which seems an odd decision
I'm not sure why it wasn't just part of the HTML 🤔). Since [2] was merged
there is no code with a data-aria-controls attribute anymore so this
module was essentially a no-op.

[1]: https://github.com/alphagov/finder-frontend/commit/2744c07ef25e2c1edb53dea2ae62ee6c5c64a774#diff-f90fee1f2d7d2b2341b403e10727271b70bd61e365b4ea19e7df1e9fcf789e19
[2]: https://github.com/alphagov/finder-frontend/commit/af5b72012911bc2d84ec4af86811f5109d063d6a#diff-f90fee1f2d7d2b2341b403e10727271b70bd61e365b4ea19e7df1e9fcf789e19