Convert track-brexit-qa-choices to GOV.UK approach
This converts this module to use the prototype class based approach to
defining a module class. This is the more common convention for GOV.UK
frontend code.

It also adjusts the test code to execute the testing on the module
element itself rather than the parent element. This was different to how
the module is run in the app and the module code is not compatible with
this.