Replace google API gem
The [google-api-client gem has been deprecated](https://github.com/googleapis/google-api-ruby-client/blob/master/google-api-client/OVERVIEW.md)
and is no longer receiving updates. It was a rollup gem containing a heap of
dependencies, and was getting pretty big.

We can replace it with the gems for the individual services we require
(which we have now done!)

Output from old gem:

```
*******************************************************************************
The google-api-client gem is deprecated and will likely not be updated further.

Instead, please install the gem corresponding to the specific service to use.
For example, to use the Google Drive V3 client, install google-apis-drive_v3.
For more information, see the FAQ in the OVERVIEW.md file or the YARD docs.
*******************************************************************************
```