Return A/B response headers for all variants
The `Vary` header is only be returned for the "B" variant which makes it
tricky for Fastly to know how to cache the pages correctly. This has
been updated to always pass the `Vary` header back to Fastly regardless
of the variant.

Tests have been updated to check that this is working correctly, with
checks for the Vary header, the meta tag and the correct Slimmer
template header.