Set Transition Checker to use old style header
We will be setting the Explore Super Menu Header as the default across GOV.UK, so we need to switch the Brexit landing page to use a specific template from Static that has a header with the Sign in link. This is temporary, while we work on design iterations to add the Accounts related links.