Disable GOV.UK account creation during migration
We're planning to migrate the GOV.UK account to use Digital Identity's
auth service. During this time users will be unable to create new GOV.UK
accounts, so temporarily remove the button and replace it with a warning.

Once the migration is complete we can revert this commit and make any
further changes needed for the switchover.