Remove email signups links on checker results page
As we're retiring the checker in 3 weeks and not intending to make any
more updates which would result in emails being sent, we should remove
the links to sign up.

There's a fair amount of stuff remaining for accounts users. I'm intending
to leave that in place until we actually retire the checker, as it makes
things simpler here, and reduces the risk of us breaking things ahead of
time.

This does mean that we have to fake a couple of navigation steps in accounts
tests so that the remaining steps can run.