Restore full width blue header
Because the blue header is hidden behind a query string (for example,
`?parent=%2Fbrexit&topic=d6c2de5d-ef90-45d1-82d4-5f2438369eea`) it's easily
missed.

This changes the template requested from Static to the full width version.
The search results are already contained inside an element with a
`.width-container` class[1], so this doesn't affect rest of the the page's
layout.

[1]: https://github.com/alphagov/finder-frontend/blob/e83059d079691e2b37b2e1da761b4155c854adac/app/views/finders/show.html.erb#L37