Remove some Brexit checker code
We've broken these tests by adding redirects away from checker routes in
the previous commit.

This removes some of the code used by the Brexit checker. The rest will be
tidied up completely once we're happy. We need to retain some of it in case
we need to send another email to update users of further changes.