Remove cache key checks from readiness probe
This will enable Finder Frontend to serve requests before
its cache is warm. The readiness probe should only check
the hard dependencies of the application (ie are
required to serve all requests). This allows
the application to degrade gracefully when external
dependencies fails, and prevents a cascading failure
when those dependencies fail.