Remove file extension
Having an extension in a partial name was deprecated in Rails 6.1 and
removed in Rails 7 (https://github.com/rails/rails/pull/39164)

The file extension parameter passed to the render_as_string needs to
be a symbol. I suspect this wasn't working before, but it didn't matter because
the file name contained the format type.