Create `/search/latest` endpoint
We are retiring `/government/latest` from being rendered in Whitehall.

Finder Frontend's `/search/all` finder provides the same search
response, so we are redirecting `/government/latest` to this finder.

However there is a mismatch in the query string parameters used between
the two applications. Therefore we need to redirect any queries
containing `departments` to match the expectation of the finder
(`organisations`). Additionally, we need to add a default search order
to see the latest documents instead of the most relevant documents.