Test we don't have JS vulnerabilities in headless browser
This imports a test from Smokey that confirmed Finder Frontend doesn't
have execute JavaScript that is inputted by a user [1]. This has been
ported because this testing can and should be localised to this
application and doesn't need to be done in a wider, distinct suite.

It does seem unusual to have a test of this sort of low-level concern at
a feature level, because it is not a feature. However given we've had
at-least 2 incidents related to XSS in Finder Frontend [2], [3] it seems
like something we should check. As a feature test operates in a real
browser environment this is the only place we can have high confidence
that input doesn't cause unexpected JavaScript execution.

This test doesn't check the server side handling of XSS a separate
change is being added for that.

In order to make these tests work and not had a whole bunch of new steps
I had to do some wrestling with the step definitions. They don't seem to
be in a good state, so I just tried to do minimal damage and keep it all
running.

[1]: https://github.com/alphagov/smokey/blob/7553a432259d785fb6d4fb81cfd4fe3572185ad3/features/apps/finder_frontend.feature#L11-L31
[2]: https://github.com/alphagov/finder-frontend/pull/480
[3]: https://github.com/alphagov/finder-frontend/pull/1060