Test for XSS vulnerabilities at a server side level
In the previous commit I added a test that checked in a real browser
environment that XSS issues didn't occur when using finder frontend with
JS.

This adds a more basic test that handles things at a basic HTTP
request/response level which asserts that if someone enters JS in a
search query it doesn't get output on the page.

As per the previous commit, this very cautious approach to XSS is based
off two previous incidents with XSS which have led to us having quite
expensive smoke tests constantly checking for XSS [1] [2]. Adding this allows us
to replace those.

[1]: https://github.com/alphagov/smokey/blob/7553a432259d785fb6d4fb81cfd4fe3572185ad3/features/apps/finder_frontend.feature#L11-L31
[2]: https://github.com/alphagov/smokey/blob/7553a432259d785fb6d4fb81cfd4fe3572185ad3/features/apps/finder_frontend.feature#L61-L77