Don't error 500 on invalid params
This is to fix an issue where we're seeing error alerts in Sentry due to
a client input issue.

A client is requesting:

/search/latest?departments%5B%5D/government/organisations/department-for-business-energy-and-industrial-strategy&page=1

(note the lack of equals following `departments%5B%5D`)

This causes Rails to misinterpret the request and raise an exception of: ActionController::UnfilteredParameters

This commit changes the behaviour of the code so that it will only
accept input that is in an array format. Thus if a user inputs anything
strange the parameter will be rejected and thus the server won't error.