Turn off font
With the font being served by each application, users need to re-download it
every time they went from a page rendered in one application to a page
rendered in another application.

Serving the font files from Static - instead of each application - means that
they're cached across the entirety of GOVUK; so the font shouldn't be included
in Finder-frontend's stylesheet.