Fix OIM capitialisation issue
The main issue appears to be that ‘facet.key.humanize’ is being applied to a hash key (e.g. ‘oim_project_type’), whereas the users simply want to see what they have added to the content item - in this case ‘OIM Project type’.

Using the ‘format’ field from the content item is not ideal. However, the tests indicate that ‘facet.key.humanize’ works for other documents. I question if we should just be using ‘facet.name’ in all cases.