Fix accessibility issue with mobile filters
Our mobile filters was previously a popup/modal. The issue with this was that Android TalkBack and iOS VoiceOver's focus would not get trapped inside the popup. Therefore, users would hear about elements behind the popup. As well as this, Safari would give keyborad focus to elements behind the modal while it was open. Therefore, the mobile filters section has been changed to be an expandable in-page element.