Implement announcement finder redirects
These redirects used to be rendered by Whitehall, but were removed in
https://github.com/alphagov/whitehall/pull/6950.

As Whitehall Frontend is being retired, we need to bring these into
another application.  The previous routes are already being redirect to
Finder Frontend, so are best processed here.

This replicates the behaviour in an identical way to Whitehall, see the
code before it was removed:
https://github.com/alphagov/whitehall/blob/ee7bd0d4caa90dd5496158da292b8cdc5c5f4b39/app/controllers/announcements_controller.rb