Replace inline data image with file asset
This replaces the usage of an inline data image from a CSS file. The
motivation for doing this is the Content Security Policy removing the
ability to embed data images for improved security [1].

Unfortunately this duplicates an asset that is in the component gem [2],
it felt fragile and delicate to reference the asset directly from
publishing components. Ideally we'd actually use the component from the
gem, but this seemed impractical to achieve just now (mostly because of
extra margins).

[1]: https://github.com/alphagov/govuk_app_config/pull/279
[2]: https://github.com/alphagov/govuk_publishing_components/blob/main/app/assets/images/govuk_publishing_components/icon-search.svg