Update caching strategy for pption select fact view
When the facet presentation order changes, but the actual content remains otherwise
unchanged, this does not invalidate the cache for this view fragment.

Updating this to include the OptionSelectFacet's version in the cache key ensures
that the view if always updated when the underlying facet's model changes in anyway.
not always