Add option to remove top border on first list item
On specialist finder pages (or finder pages where no sort options are rendered) a double border is seen since no sort options are displayed. The border on the first document list item should be removed. On the main search results page the top border on the first document list item is necessary to visually separate the sort options from the document list.