Add support for `open_on_load` facet option
By default the option select facet will be closed if it has no selected items and it is not the first option select facet on the page.

If the `open_on_load` facet option is provided and set to true, it will override the default logic and ensure the option select facet is open.