Disable LTR ranking on licence transaction path.
Use hardcoded utility function introducted to catch boosts
and also turn off LTR ranking using ab_tests=disable:relevance
on that path, which improves search results on the licence
finder.

We're using the ab parameters relevance:disable to do this.
Since this isn't strictly an A/B thing it feels a bit wrong to
use this parameter, but it's all that's available to us at the
moment. We'll add a tech debt card to examine making this a
proper parameter.