Add data-ga4-change-category and Ga4FinderTracker call to finders
This allows us to track "change" events on the Finders, so that we can push GA4 events when the keyword or filters change.

We add data-ga4-section so that we know the element is a filter section. This also gives us the title of the section for our GA4 object. We add data-ga4-filter-container so that our JS knows where to hook into to setthe indexes. It will loop through the data-ga4-section elements within this container to set the indexes. We initialise the setFilterIndexes function when the search page loads.