[DO NOT MERGE] Adapt app for prototype deployment
We are building a series of prototype APIs to test integration with
several candidate search engine products. This adapts the application to
be used as a frontend for these prototype APIs by:
- Adding a conspicuous prototype warning banner
- Adding a .cfignore file and changing the Procfile to allow for GOV.UK
  PaaS deployment

The intention is that this branch of the app will be deployed. This code
should not be merged into `main`.