Remove the google-apis-drive_v3 gem
I came across this gem during a dependabot update and upon inspection, it doesn't appear to be used anymore. It seems like it was added to convert an actions CSV into YAML [1] and was used by the Brexit checker [2]. There was a related method in a fixtures file in the main branch but it wasn't being called anywhere. The unused methods and gem have therefore been removed in this PR.

[1](https://github.com/alphagov/finder-frontend/pull/1485/files)
[2](https://github.com/alphagov/finder-frontend/tree/db8d28aaa06642aeab4b2f91842f10c974a6a4c0/app/lib/brexit_checker)