Remove feed signup link if licence finder
The plan is to remove the feed links from all finders, unless we gain
evidence that it is needed. Until this decision is made, we don't want to
introduce the feature for the new licence finder.

A new generalised method could be implented (not use
`is_licence_transaction?`) if we wanted to do this for a subset of
finders, vs all.