Remove AA test config
Removing this code as the AA test has been run.

There is however a chance that we will run a throttled version of the same test in a few weeks...
Given that it is being removed from the CDN config, I also decided to fully
remove the configuration from here, as it will be quick enough to bring it back again in
a few weeks if required.

PR to implment the AA test config: https://github.com/alphagov/finder-frontend/pull/3053