Update AA test config for ElasticSearch 6.7
There is currently an AA test running with a 50/50 split [1].
We now want to run a throttled version with a 0.5% split.
The C variant has therefore been added to the corresponding
govuk-cdn-config PR to soak up the traffic not in A or B.
This PR allows the C variant. Z will remain as the control variant.
[1]https://github.com/alphagov/finder-frontend/pull/3135