Rename classes local in option_select partial
This array of CSS classes is actually applied to the options container
within the component and not to the top-level element.

Also govuk_publishing_components components typically allow the CSS
classes on the top-level element to be customised via a `classes` local
variable, so I think the new name is less confusing.